requirements.txt
fastapi
uvicorn
dspy-ai==2.4.5
chromadb
pydantic
google-generativeai
llama_index
llama-index-embeddings-gemini==0.1.6
